xen: platforms: Correct pm_rst_access to map a node into a memory address
EEMI call for RESET_ASSERT calls pm_check_access() to check if a domain is
permitted to access particular iomem. Previously, this check was failing because
pm_rst_access mapped a reset node to corresponding device node instead of the
device address and pm_check_access() was unable to find the iomem regions.

Signed-off-by: Vikram Garhwal <fnu.vikram@xilinx.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@xilinx.com>