libxl: default nic type should be vif
The commit ("libs: light: Add virtio-net support") changed the default vif type
to ioemu. As a consequence, if the user does not specify the nic type in the xl
config file, domU creation would fail due to incorrect device model defaults. To
successfully create a domU, it was required to specify type=vif in the xl config
file:

vif = [ 'type=vif' ]

However, omitting the type is a common use case. The default type should be vif,
not ioemu. This patch sets the default nic type to vif again. We can now again
create domUs with a regular vif by omitting the type:

vif = [ '' ]

Or, to create a virtio-net device:

vif = [ 'model=virtio-net,type=ioemu' ]

Fixes: ("libs: light: Add virtio-net support")
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
Reviewed-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>