xen/eemi: Do not forward PM_CLOCK_DISABLE to firmware
At the moment, in order to prevent Linux guests to disable the clocks that
are not required by any device, we need to specify clk_ignore_unused
cmdline option. Otherwise, one domain could disable the clocks that might
be in use by another domain, potentially causing issues if the latter is
already utilizing a device or if the firmware + clock controller node
have not been passthroughed.

To accommodate the recent requests to get rid of this property, modify
the handling so that PM_CLOCK_DISABLE calls are ignored by returning
success without forwarding the PM API call to firmware.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>