x86/hvm: fix dom0 legacy irqs by advertising PCAT compat
Advertise PCAT compatibility so that dom0 proceeds with enabling
and using legacy interrupts (ex. rtc irq 8, gpio irq 7).

Since PCAT compatibility indicates the presence of a legacy PIC,
Linux Dom0 will attempt to configure PIT, will fail and crash.
To prevent this, enable PIT emulation for Dom0.

IRQ2 is either cascade IRQ or is used for IRQ0 override,
hence it is handled by Xen and cannot be passed through
to Dom0. So when Dom0 attempts to set it up in IO-APIC,
skip identity mapping, and let its source to be the
virtual platform timer as is for unprivileged domains.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Change-Id: Ic119bfd223517c3a7ff5ae1b9b114dd2a41dc29e