x86/pvh: emulate C2 state to support s2idle in the guest
The Processor Register Block’s (P_BLK’s) P_LVL2 register coupled with
FADT P_LVL2_LAT values provide support for placing the system processors
into the C2 state. System software reads the P_LVL2 register to enter
the C2 state. The hardware must always exit this state whenever an
interrupt is to be presented to the processor. The P_BLK address is
defined in DSDT Processor block.

In order to emulate C2 state in xen, the reads of the P_LVL2 register
are trapped and the trapped vcpu is blocked. The blocked vcpu will
be rescheduled when a event is to presented to it.

The function vcpu_block() was open-coded in the enter_c2() handler,
to check explicitly for evtchn upcall events, because there was
a situation where the vcpu was blocked while there was a pending
upcall due to a race. The upcall event was not registered yet in
vlapic and that lead the vcpu to get blocked after being kicked.

Signed-off-by: Xenia Ragiadakou <xenia.ragiadakou@amd.com>
Change-Id: I2168b8b04717f010c151e0efe02fbed7ce99e937