xen/arm: Add way to disable traps on accesses to unmapped addresses
Add a per-domain way to optionally disable traps for accesses
to unmapped addresses.

The domain flag is general but it's only implemented for Arm for now.

For the backport into the Xilinx tree, I've kept the upstream
definition of trap_unmapped_accesses at bit 8 to keep things
aligned with upstream, i.e:

+#define XEN_DOMCTL_CDF_trap_unmapped_accesses  (1U << 8)
-#define XEN_DOMCTL_CDF_vpci           (1U << 8)
+#define XEN_DOMCTL_CDF_vpci           (1U << 9)

Signed-off-by: Edgar E. Iglesias <edgar.iglesias@amd.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>